% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{approxDeriltau}
\alias{approxDeriltau}
\title{Approximate derivative w.r.t. \eqn{\log(\tau)}}
\usage{
approxDeriltau(ltau, minibatchSize, k)
}
\arguments{
\item{ltau}{\eqn{\log(\tau)}.}

\item{minibatchSize}{Minibatch size.}

\item{k}{The latent dimension.}
}
\description{
\code{approxDeriltau} calculates derivative w.r.t. \eqn{\log(\tau)}.
}
