% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{approxDeriTransformedTau}
\alias{approxDeriTransformedTau}
\title{Approximate derivative w.r.t.
\eqn{\text{logit}\{(\tau - \text{lowerB}) / (\text{upperB} - \text{lowerB})\}.}}
\usage{
approxDeriTransformedTau(
  transformedPara,
  lowerB,
  upperB,
  minibatchSize,
  k,
  fac
)
}
\arguments{
\item{transformedPara}{\eqn{\text{logit}\{(\tau - \text{lowerB}) / (\text{upperB} - \text{lowerB})\}}.}

\item{lowerB}{The lower bound of tau.}

\item{upperB}{The upper bound of tau.}

\item{minibatchSize}{The minibatch size.}

\item{k}{The latent dimension.}

\item{fac}{Default is 1.}
}
\description{
\code{approxDeriTransformedTau} approximates the derivative of \code{\link[=calclogC_up2const]{calclogC_up2const()}}
w.r.t.  \eqn{\text{logit}\{(\tau - \text{lowerB}) / (\text{upperB} - \text{lowerB})\}.}
}
