% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaMatMult}
\alias{armaMatMult}
\title{Matrix multiplication}
\usage{
armaMatMult(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\description{
\code{armaMatMult} calculates the matrix product of A and B.
}
