% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{assignSamples}
\alias{assignSamples}
\title{Assign samples}
\usage{
assignSamples(sampleList, rowIdx, cur_sample, par_name)
}
\arguments{
\item{sampleList}{A list of samples.}

\item{rowIdx}{The row index.}

\item{cur_sample}{A list of samples from the current iteration.}

\item{par_name}{Parameter names.}
}
\value{
A list of samples
}
\description{
\code{assignSamples} assign the sample from current iteration to a list of samples.
}
