% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_lambda_grad_cpp}
\alias{calc_lambda_grad_cpp}
\title{Gradient of the log-likelihood w.r.t. lambda}
\usage{
calc_lambda_grad_cpp(lambda, psdMat_grad, vec_grad, invSigma, centered_sdat)
}
\arguments{
\item{lambda}{A numeric matrix, p by k, the factor loading matrix.}

\item{psdMat_grad}{The gradient w.r.t. the matrix parameter in the pseudo-normalising
constant.}

\item{vec_grad}{The gradient w.r.t. the vector parameter in the pseudo-normalising
constant.}

\item{invSigma}{A vector of precisions, length = p.}

\item{centered_sdat}{A matrix of centered mini-batch data, n by p.}
}
\description{
\code{calc_lambda_grad_cpp} calculates the gradient of the log-likelihood w.r.t.
lambda, where lambda is the factor loading matrix.
}
