% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_lambda_grad_MatPart2_cpp}
\alias{calc_lambda_grad_MatPart2_cpp}
\title{Gradient for the factor loadings}
\usage{
calc_lambda_grad_MatPart2_cpp(lambda, psdMat_grad, invSigma)
}
\arguments{
\item{lambda}{A matrix, the factor loading.}

\item{psdMat_grad}{A matrix, the gradient w.r.t. the positive semi-definite matrix.}

\item{invSigma}{A vector, the precisions.}
}
\description{
\code{calc_lambda_grad_MatPart2_cpp} calculates the gradient of the log-likelihood through
the matrix parameter in the pseudo-normalising constant w.r.t. the factor loadings matrix.
}
