% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_lambda_grad_VecPart_cpp}
\alias{calc_lambda_grad_VecPart_cpp}
\title{Gradient for the factor loadings}
\usage{
calc_lambda_grad_VecPart_cpp(invSigma, centered_sdat, vec_grad)
}
\arguments{
\item{invSigma}{A numeric vector, the precisions.}

\item{centered_sdat}{A numeric matrix, centered minibatch.}

\item{vec_grad}{A vector, the gradient of the log-likelihiood
w.r.t. the vector parameter.}
}
\description{
\code{calc_lambda_grad_VecPart_cpp} calculates the gradient of the log-likelihood
through the vector parameter in the pseudo-normalising constant w.r.t.
the factor loadings matrix.
}
