% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_linvSig_grad_cpp}
\alias{calc_linvSig_grad_cpp}
\title{Gradient for the precision parameters}
\usage{
calc_linvSig_grad_cpp(invSig, psdMat_grad, vec_grad, lambda, centered_sdat)
}
\arguments{
\item{invSig}{The precision parameters.}

\item{psdMat_grad}{The gradient w.r.t. the matrix parameter in the pseudo-normalising
constant.}

\item{vec_grad}{The gradient w.r.t. the vector parameter in the pseudo-normalising
constant.}

\item{lambda}{A numeric matrix, the factor loading matrix.}

\item{centered_sdat}{A numeric matrix, centered minibatch data.}
}
\description{
\code{calc_linvSig_grad_cpp} calculates the gradient of the log-likelihood w.r.t.
the function of the precisions that is being updated in the sampling procedure.
}
