% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{calc_transTau_grad}
\alias{calc_transTau_grad}
\title{Derivative w.r.t the transformed tau parameter}
\usage{
calc_transTau_grad(tau, vec_grad_sum, mu, tauBounds)
}
\arguments{
\item{tau}{The value of tau.}

\item{vec_grad_sum}{XXXX}

\item{mu}{A vector of length k.}

\item{tauBounds}{A list of tau bounds, with names \code{lowerB} and
\code{upperB}.}
}
\value{
The log-likelihood derivative w.r.t tau
}
\description{
\code{calc_transTau_grad} calculates the derivative w.r.t. the transformed tau
parameter, namely the parameter that is being directly updated in the sampling
procedure.
}
