% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{calcDericenter}
\alias{calcDericenter}
\title{Calculate derivative for parameters with a normal prior}
\usage{
calcDericenter(invSig1, centered_sdat)
}
\arguments{
\item{invSig1}{The inverse of noise variances, precision.}

\item{centered_sdat}{Centered minibatch, a numeric matrix.}
}
\description{
\code{calcDericenter} calculates derivative w.r.t parameters (e.g. center and )
}
