% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{calcDerilinvSig}
\alias{calcDerilinvSig}
\title{Calculate derivative w.r.t \eqn{\log(1 / \sigma^2)}}
\usage{
calcDerilinvSig(invSig1, centered_sdat)
}
\arguments{
\item{invSig1}{The precision.}

\item{centered_sdat}{Centered minibatch, a numeric matrix.}
}
\value{
A vector, gradient w.r.t \eqn{\log(1 / \sigma^2)}.
}
\description{
\code{calcDerilinvSig} calculates the derivative w.r.t. \eqn{\log(1 / \sigma^2)},
where \eqn{\sigma^2} follows a normal prior.
}
