% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{calcFB_pars}
\alias{calcFB_pars}
\title{Fisher-Bingham parameters}
\usage{
calcFB_pars(invSigma, lambda, tau, mu, center, dat, norm_)
}
\arguments{
\item{invSigma}{The precision.}

\item{lambda}{A matrix, p by k, factor loadings.}

\item{tau}{The concentration parameter.}

\item{mu}{A vector of length k.}

\item{center}{A vector of length p.}

\item{dat}{A numeric matrix.}

\item{norm_}{Logical, whether normalized \code{para1} or not.}
}
\description{
\code{calcFB_pars} calculates the parameters for the Fisher-Bingham distribution
that the latent factors follow a posteriori.
}
