% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{calcGrad_llh_fast}
\alias{calcGrad_llh_fast}
\title{Calculates the gradient of the log-likelihood w.r.t each parameter}
\usage{
calcGrad_llh_fast(
  par_names,
  invSigma,
  lambda,
  tau,
  mu,
  centered_sdat,
  tauBounds,
  accurate_grad
)
}
\arguments{
\item{par_names}{A list of parameter names.}

\item{invSigma}{A vector of the noise precision.}

\item{lambda}{The factor loading matrix, p by k.}

\item{tau}{The concentration parameter}

\item{mu}{The direction parameter, a vector of length k.}

\item{centered_sdat}{A matrix of centered data that are in the minibatch.}

\item{tauBounds}{The lower and upper bounds of tau}

\item{accurate_grad}{Logical, whether accurate gradient is desired or not.}
}
\value{
A list of gradients for the parameters in \code{par_names}.
}
\description{
\code{calcGrad_llh_fast}is a high-level function that calculates the gradient of the log-
likelihood w.r.t parameters that are in \code{par_names}.
}
