% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{calcGrad_llh_fast_wrapper}
\alias{calcGrad_llh_fast_wrapper}
\title{Wrapper of \code{\link[=calcGrad_llh_fast]{calcGrad_llh_fast()}}}
\usage{
calcGrad_llh_fast_wrapper(par_names, theta1, sdat, tauBounds, accurate_grad)
}
\arguments{
\item{par_names}{Parameter names.}

\item{theta1}{The parameter values.}

\item{sdat}{The minibatch.}

\item{tauBounds}{A list of bounds for tau, lowerB and upperB.}

\item{accurate_grad}{Logical, whether accurate gradient is desired.}
}
\description{
\code{calcGrad_llh_fast_wrapper} is wrapper to \code{\link[=calcGrad_llh_fast]{calcGrad_llh_fast()}}.
}
