% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcgradLambdaprior_cpp}
\alias{calcgradLambdaprior_cpp}
\title{Prior gradient of Lambda}
\usage{
calcgradLambdaprior_cpp(Lambda, phi, kappa)
}
\arguments{
\item{Lambda}{A numeric matrix, the factor loading matrix.}

\item{phi}{A prior parameter in the Dirichlet-Laplace prior.}

\item{kappa}{A prior parameter in the Dirichlet-Laplace prior.}
}
\description{
\code{calcgradLambdaprior_cpp} calculates the gradient of the Dirichlet-Laplace
prior w.r.t. Lambda,  a factor loading matrix.
}
