% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{calcgradTau_truncatedNormal_prior}
\alias{calcgradTau_truncatedNormal_prior}
\title{Derivative w.r.t. the function of interest of tau}
\usage{
calcgradTau_truncatedNormal_prior(ltau, prior_par)
}
\arguments{
\item{ltau}{The parameter that is being updated in the sampling procedure.}

\item{prior_par}{A list of prior parameters, ordered by names \code{lowerB},
\code{upperB}, \code{cvmf} (mean), \code{dvmf} (sd).}
}
\description{
\code{calcgradTau_truncatedNormal_prior} calculates derivative w.r.t the function of tau,
that is actually being updated in the sampler, with tau following a truncated
normal prior \eqn{\text{N}(\text{mean}, \text{sd}^2)\mathbb{I}(l,u).}
}
\details{
The parameter that is being updated in the sampler is
\eqn{\text{logit}\{(\tau - l)/(u - l)\}.}
}
