% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{calcllhgrad}
\alias{calcllhgrad}
\title{Gradient of the log-likelihood}
\usage{
calcllhgrad(par_names, theta1, sdat, k, tauBounds, accurate_grad)
}
\arguments{
\item{par_names}{A list of parameters to be updated}

\item{theta1}{A list of parameter values}

\item{sdat}{The minibatch data.}

\item{k}{The latent dimension.}

\item{tauBounds}{A list of tau information, lowerB, upperB and fac.}

\item{accurate_grad}{Logical, whether accurate or simple gradient is desired.}
}
\description{
\code{calcllhgrad} calculates gradient of log-likelihood w.r.t. the parameters in
\code{par_names}.
}
