% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_density_eg.R
\name{calclogC_up2const}
\alias{calclogC_up2const}
\title{The logarithm of the normalizing constant in the von Mises-Fisher distribution}
\usage{
calclogC_up2const(tau, k, logged)
}
\arguments{
\item{tau}{The concentration parameter (When on the original scale, tau > 0).}

\item{k}{The dimension.}

\item{logged}{Whether \eqn{\tau} is logged or not.}
}
\value{
The normalizing constant on the logarithm scale.
}
\description{
\code{calclogC_up2const} returns the logarithm of the normalising constant in
the von Mises-Fisher distribution up to a constant (i.e. excluding \eqn{(2 \pi)^{k / 2}}).
}
\details{
\code{calclogC_up2const} calculates the logarithm of the normalizing constant in
a von Mises-Fisher distribution up to a constant. The quantity is defined by
\deqn{ C_k(\tau) = \frac{\tau^{k / 2 - 1}}{ I_{k / 2 - 1}(\tau)}.}
}
\examples{
calclogC_up2const(5, 3, TRUE)
}
