% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{calcloglaplaceDeri}
\alias{calcloglaplaceDeri}
\title{Derivative of Laplace prior}
\usage{
calcloglaplaceDeri(var, scale, mean = 0)
}
\arguments{
\item{var}{The variable.}

\item{scale}{Scale of the distribution.}

\item{mean}{Mean of the distribution.}
}
\description{
\code{calcloglaplaceDeri} calculates the derivative w.r.t. \code{var}.
}
