% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calclogPseudoconst_MatParGrad4_cpp}
\alias{calclogPseudoconst_MatParGrad4_cpp}
\title{Gradient w.r.t. the matrix parameter in the pseudo-normalising constant}
\usage{
calclogPseudoconst_MatParGrad4_cpp(para2_vech, para1)
}
\arguments{
\item{para2_vech}{A vector obtained through vech operation from the matrix
parameter.}

\item{para1}{A vector, the vector parameter in the pseudo-normalising constant.}
}
\description{
\code{calclogPseudoconst_MatParGrad4_cpp} calculates the gradient of the logarithm
of the pseudo-normalising constant w.r.t. the matrix parameter.
}
