% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{calclogPseudoconst_VecParGrad_fast}
\alias{calclogPseudoconst_VecParGrad_fast}
\title{Gradient w.r.t the first parameter of a pseudo-normalising constant}
\usage{
calclogPseudoconst_VecParGrad_fast(
  vec_par,
  eigVals,
  eigVecs,
  ord_eigVal,
  order_ = 3
)
}
\arguments{
\item{vec_par}{The vector parameter value, a vector of length k.}

\item{eigVals}{Eigenvalues of the matrix parameter, a vector of length k.}

\item{eigVecs}{Eigenvectors of the matrix parameter, a matrix.}

\item{ord_eigVal}{XXX}

\item{order_}{The order of saddlepoint approximation, 1: first-order, 2:
second-order first version, 3: second-order second version. Default is 3 for
better accuracy. See Kume and Wood (2005) P975 for more details.}
}
\description{
\code{calclogPseudoconst_VecParGrad_fast} calculates the gradient of the
log-likelihood w.r.t. the vector parameter of the pseudo-normalising constant
}
\references{
Kume, A., & Wood, A. T. A. (2005). Saddlepoint Approximations for the
Bingham and Fisher-Bingham Normalising Constants. Biometrika, 92(2), 465–476.
http://www.jstor.org/stable/20441200
}
