% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{calcpriorgrad}
\alias{calcpriorgrad}
\title{Prior gradient}
\usage{
calcpriorgrad(theta1, p, k, prior_para)
}
\arguments{
\item{theta1}{A list of the parameter values, ordered by their names}

\item{p}{The ambient dimension}

\item{k}{The latent dimension}

\item{prior_para}{A list of the prior parameters}
}
\description{
\code{calcpriorgrad} computes the prior gradient w.r.t the parameters in \code{theta1}.
}
