% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcPsdMat_linvSig_jaco_cpp}
\alias{calcPsdMat_linvSig_jaco_cpp}
\title{Gradient for the precision parameters}
\usage{
calcPsdMat_linvSig_jaco_cpp(invSig, psdMat_grad, lambda)
}
\arguments{
\item{invSig}{A numeric vector, the precision parameters.}

\item{psdMat_grad}{The Jacobian w.r.t. the positive semi-definite matrix.}

\item{lambda}{A numeric matrix of p by k, the factor loading matrix.}
}
\description{
\code{calcPsdMat_lingSiv_jaco_cpp} calculates the Jacobian of the pseudo-normalising constant
w.r.t. \code{invSig} through change in the positive semi-definite matrix parameter.
}
