% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{calcSGgrad}
\alias{calcSGgrad}
\title{Computes stochastic gradient}
\usage{
calcSGgrad(
  theta1,
  dat,
  k,
  prior_para,
  accurate_grad,
  minibatchSize,
  initialization_period,
  prior_exclude
)
}
\arguments{
\item{theta1}{A list of parameter values at which gradient is evaluated.}

\item{dat}{The minibatch, a numeric matrix of n by p}

\item{k}{The latent dimension.}

\item{prior_para}{A list of prior parameters.}

\item{accurate_grad}{Logical, whether to evaluate the accurate gradient.}

\item{minibatchSize}{Minibatch size.}

\item{initialization_period}{Logical, whether it is the initialization period.}

\item{prior_exclude}{Logical, whether to ignore the priors in evaluating
the gradient.}
}
\description{
\code{calcSGgrad} computes the stochastic gradient for \code{dat} at parameter
values \code{theta1}.
}
