% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcVec_linvSig_jaco_cpp}
\alias{calcVec_linvSig_jaco_cpp}
\title{Gradient for the precision parameters}
\usage{
calcVec_linvSig_jaco_cpp(invSig, vec_grad, lambda, centered_sdat)
}
\arguments{
\item{invSig}{A numeric vector the precision parameters.}

\item{vec_grad}{The gradient of the pseudo-normalising constant w.r.t. the vector.}

\item{lambda}{A numeric matrix, the factor loading matrix.}

\item{centered_sdat}{A numeric matrix, centered minibatch.}
}
\description{
\code{calcVec_linvSig_jaco_cpp} calculates the Jacobian of the pseudo-normalising constant
w.r.t. \code{invSig} through change in the vector parameter.
}
