% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{calcVecGrad_all}
\alias{calcVecGrad_all}
\title{Gradient w.r.t the first parameter in the pseudo-normalising constant}
\usage{
calcVecGrad_all(para1, ord_eigVals, eigVecs, ord, keyword)
}
\arguments{
\item{para1}{The first parameter in the pseudo-normalising constant,
a matrix of k by minibatch size.}

\item{ord_eigVals}{XXXX}

\item{eigVecs}{The eigenvectors of the second (or matrix) parameter in the
pseudo-normalising constant}

\item{ord}{The order of saddlepoint approximation, 1: first-order, 2:
second-order first version, 3: second-order second version. Default is 3 for
better accuracy. See \insertCite{KumeWood05}{ellipsoidgaussian}}

\item{keyword}{'Richardson' or 'simple', the method used for gradient
approximation.}
}
\value{
The gradient w.r.t. the first parameter in the pseudo-normalising
constant.
}
\description{
\code{calcVecGrad_all} calculates the gradient w.r.t. the first parameter in the
pseudo-normalising constant for all the data in the minibatch.
}
