% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{collectSamples}
\alias{collectSamples}
\title{Save the samples}
\usage{
collectSamples(sampleList, notUpdatedList, nrow, tauBounds)
}
\arguments{
\item{sampleList}{A list of the raw samples from the gSGNHT sampling}

\item{notUpdatedList}{A list of the parameters that are not being updated in the
sampling.}

\item{nrow}{Number of rows}

\item{tauBounds}{A list of tau bounds, \code{lowerB} and \code{upperB}.}
}
\value{
A list of organized samples, on the scale of interest.
}
\description{
\code{collectSamples} transforms the updated parameters to their original scales
and saves them in a desired format.
}
