% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{compute_stiefel_flow}
\alias{compute_stiefel_flow}
\title{Compute stiefel flow}
\usage{
compute_stiefel_flow(X, V, h)
}
\arguments{
\item{X}{The orthogonal matrix, a p by k matrix \eqn{p \geq k}.}

\item{V}{A p by k matrix \eqn{p \geq k}.}

\item{h}{step size}
}
\description{
\code{compute_stiefel_flow} computes the stiefel flow
}
