% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{cor_func}
\alias{cor_func}
\title{Correlation for the scatter plot matrix}
\usage{
cor_func(data, mapping, method, symbol, ...)
}
\arguments{
\item{data}{The data set to be plotted.}

\item{mapping}{The mapping.}

\item{method}{The method of correlation calculation.}

\item{symbol}{Symbol.}

\item{...}{...}
}
\description{
\code{cor_func} calculates correlation used for the scatter plot matrix
constructed by GGally::ggpairs().
}
