% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\docType{data}
\name{ctef}
\alias{ctef}
\title{global reference to scipy}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
ctef
}
\description{
@description
\code{ctef} will be initialized in .onLoad.
}
\keyword{datasets}
