% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_density_eg.R
\name{dellipsoidgaussian_up2const_logged}
\alias{dellipsoidgaussian_up2const_logged}
\title{Log density up to a constant}
\usage{
dellipsoidgaussian_up2const_logged(dat, Sigma, lambda, tau, mu, center)
}
\arguments{
\item{dat}{Data in a matrix format (no. observations x no. dimensions p).}

\item{Sigma}{The noise variances, a vector of length p.}

\item{lambda}{The factor loading matrix, \eqn{p \times k}, where k is the number of latent dimensions.}

\item{tau}{The concentration parameter (> 0).}

\item{mu}{The mean direction parameter, a vector of length k.}

\item{center}{The ellipsoid center, a vector of length p.}
}
\value{
The log-likelihood of the data up to constant.
}
\description{
\code{dellipsoidgaussian_up2const_logged} calculates the log-likelihood under an ellipsoid-Gaussian
distribution up to a constant.
}
\examples{
dellipsoidgaussian_up2const_logged(shell, rep(0.1,3), matrix(c(-0.5, -0.5, -0.7, -1.5, 1.3,
 0.1, 1.8, 2.2, -2.8),3,3),5, rep(1,3) / sqrt(3), rep(0,3))
}
