% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{dFB_cpp_log}
\alias{dFB_cpp_log}
\title{Log density of the Fisher-Bingham distribution}
\usage{
dFB_cpp_log(X, para1, para2, fb_const_part)
}
\arguments{
\item{X}{A matrix, 1 by n.}

\item{para1}{A matrix, 1 by n.}

\item{para2}{The matrix parameter, n by n.}

\item{fb_const_part}{The Fisher-Bingham constant. If NULL, the
function will compute it.}
}
\description{
\code{dFB_cpp_log} calculates log density of the Fisher-Bingham distribution.
}
