% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{dlogDtauDlogit_dlogit}
\alias{dlogDtauDlogit_dlogit}
\title{Calculate derivative}
\usage{
dlogDtauDlogit_dlogit(logit_val)
}
\arguments{
\item{logit_val}{The transformed tau gSGNHT is updating.}
}
\description{
\code{dlogDtauDlogit_dlogit} calculates log(DtauDlogit) w.r.t the
logit function.
}
