% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_setup.R
\name{.onLoad}
\alias{.onLoad}
\title{Delay load ctef module}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
\code{.onLoad} delays loading ctef module (will only be loaded when accessed via $).
}
