% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{draw_latent_factors}
\alias{draw_latent_factors}
\title{Sample latent factors}
\usage{
draw_latent_factors(dat, samples, burnin = NULL)
}
\arguments{
\item{dat}{A matrix of data, n by p.}

\item{samples}{A list of posterior samples, output of \code{\link[=ellipsoid_gaussian]{ellipsoid_gaussian()}}.}

\item{burnin}{Number of burn-in samples.}
}
\description{
\code{draw_latent_factors} simulates \eqn{\min(500,
\text{number of post burn-in samples})} latent factors for each observation from
their posterior distribution, a Fisher-Bingham distribution. Simulating from
a Fisher-Bingham is challenging and we use  \insertCite{Hoff09FB}{ellipsoidgaussian}.
}
\examples{
lat_facs <- draw_latent_factors(shell, samples, burnin = 2500)

}
\references{
\insertAllCited{}
}
