% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleLatentFac.R
\name{drawLatFac_chain}
\alias{drawLatFac_chain}
\title{Simulate the latent factors.}
\usage{
drawLatFac_chain(invsig, lambda, tau, mu, center, dat, nsample = 20)
}
\arguments{
\item{invsig}{The precision.}

\item{lambda}{A matrix, the factor loading matrix.}

\item{tau}{The concentration parameter.}

\item{mu}{A vector.}

\item{center}{A vector.}

\item{dat}{A numeric matrix.}

\item{nsample}{Number of samples to draw from the Metropolis-Hasting procedure.}
}
\description{
\code{drawLatFac_chain} simulates the latent factors in the von-Mises Fisher linear
factor model. They follow a Fisher-Bingham distribution a posteriori.
}
