% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient_helpers.R
\name{dtau_dlogit}
\alias{dtau_dlogit}
\title{Calculate derivative for tau}
\usage{
dtau_dlogit(ltau, lowerB, upperB)
}
\arguments{
\item{ltau}{The parameter that is being updated in gSGNHT, namely
logit((tau - lowerB) / (upperB - lowerB)).}

\item{lowerB}{Lower bound of tau}

\item{upperB}{Upper bound of tau}
}
\description{
\code{dtau_dlogit} computes the derivative of tau w.r.t
logit((tau - lowerB) / (upperB - lowerB)).
}
