% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Matrix multiplication}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\description{
\code{eigeMapnMatMult} calculates the matrix product of A and B.
}
