% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMatMult}
\alias{eigenMatMult}
\title{Matrix multiplication}
\usage{
eigenMatMult(A, B)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}
}
\description{
\code{eigenMatMult} calculates the matrix product of A and B.
}
