% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ellipsoid_gaussian.R
\name{ellipsoid_gaussian}
\alias{ellipsoid_gaussian}
\title{Fit an ellipsoid-Gaussian distribution}
\usage{
ellipsoid_gaussian(
  dat,
  k,
  scale_col,
  updateCenter,
  niter,
  minibatchSize = NULL,
  step_size = 1e-04,
  scalar_para = 0.1
)
}
\arguments{
\item{dat}{A numeric matrix of N by p, where N is number of observations and p is the data dimension.}

\item{k}{The latent dimension.}

\item{scale_col}{Logical. If true, \code{\link[=scale]{scale()}} is applied to the data so that data has mean 0 and standard deviation 1.}

\item{updateCenter}{Logical. If true, the parameter center in the ellipsoid-Gaussian distribution is updated throughout the sampler.}

\item{niter}{Total number of iterations of the sampler.}

\item{minibatchSize}{The minibatch size in the gSGNHT algorithm. The default is NULL, referring to use
\eqn{\max(4\%N, 50)} when \eqn{N \geq 50.}}

\item{step_size}{Step size in the gSGNHT algorithm. The default is \eqn{10^{-4}}; other common options are
\eqn{10^{-5}}, \eqn{0.5 \times 10^{-4}}, and \eqn{0.5 \times 10^{-3}}. See additional details in the Details section.}

\item{scalar_para}{The scalar parameter in the gSGNHT algorithm. The default is 0.1. See additional details in the Details section.}
}
\value{
All the samples, including the pre and post burn-in ones.
}
\description{
\code{ellipsoid_gaussian} fits an ellipsoid-Gaussian distribution (or a von-Mises Fisher linear factor model)
to the data using geodesic Stochastic Gradient Nose-Hoover Thermostats (gSGNHT)
\insertCite{liu16sgMCMC}{ellipsoidgaussian}.
}
\details{
The von-Mises Fisher linear factor model is defined by
\deqn{\mathbf{x} = \mathbf{c} + \Lambda \eta + \epsilon,\quad \epsilon \sim \text{N}_p({0}, \Sigma), \quad
\eta \sim \text{vMF}(\mu, \tau) \quad
\text{and} \quad  \Sigma = \text{diag}\left(\sigma_1^2, \ldots, \sigma_p^2\right).}
Marginalizing over the distribution of \eqn{\eta} yields density
\deqn{f(\mathbf{x}) = \frac{C_k(\tau)}{(2\pi)^{\frac{p}{2}}\prod_{i = 1}^p \sigma_i}
\exp\left\{-\frac{1}{2} (x - {c})^T \Sigma^{-1} (x - {c})\right\}
\varsigma\left\{ \tau \mu + \Lambda^T \Sigma^{-1
}(x - {c}) , \frac{\Lambda^T \Sigma^{-1}\Lambda}{2}\right\},}
where \eqn{C_k(\tau) = ({\tau}/{2})^{k / 2 - 1}\{\Gamma(k / 2)I_{k /2 - 1}(\tau)\}^{-1}} and \eqn{I_{v}(\cdot)}
denotes the modified Bessel function of the first kind of order \eqn{v}, with respect to the Lebesgue measure on \eqn{\mathbb{R}^p,}
\eqn{\varsigma(\vartheta, A)} is the normalizing constant in a Fisher-Bingham distribution with density
\deqn{\frac{1}{\varsigma(\kappa \vartheta, A)} \exp\left(\kappa \vartheta^T{y} - {y}^T A {y}\right).}
We call this distribution the Ellipsoid-Gaussian (EG).
}
\examples{
res <- ellipsoid_gaussian(shell, 3, FALSE, TRUE, 100)
}
\references{
\insertRef{liu16sgMCMC}{ellipsoidgaussian}
}
