% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{extractSamples}
\alias{extractSamples}
\title{Extract the row indices from each element}
\usage{
extractSamples(sampleList, rowIdx)
}
\arguments{
\item{sampleList}{A list of samples}

\item{rowIdx}{The row index of the elements that we want to extract}
}
\value{
A list of extract samples
}
\description{
\code{extractSamples} extracts the \code{rowIdx} from \code{sampleList}.
}
