% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findFBconst_cpp}
\alias{findFBconst_cpp}
\title{Fisher-Bingham normalising constant}
\usage{
findFBconst_cpp(gam, lam, which_, ordered)
}
\arguments{
\item{gam}{A vector}

\item{lam}{A vector}

\item{which_}{The order of approximation. 1: First order; 2: second order first
type; 3: second order second type.}

\item{ordered}{Whether gam or lam is ordered.}
}
\description{
\code{findFBconst_cpp} calculates the normalising constant in the Fisher-Bingham
distribution.
}
