% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{findParaDims}
\alias{findParaDims}
\title{Find the total number of parameters that are being updated}
\usage{
findParaDims(para_dim_list)
}
\arguments{
\item{para_dim_list}{A list of parameter dimensions}
}
\value{
The total number of parameters that are being updated
}
\description{
\code{findParaDims} find the total number of parameters that are being updated in
the sampler.
}
