% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{findParDims}
\alias{findParDims}
\title{Find parameter dimensions}
\usage{
findParDims(para_names, variable_interest)
}
\arguments{
\item{para_names}{The names of the parameters}

\item{variable_interest}{A list}
}
\value{
A list of parameter dimensions
}
\description{
\code{findParDims} finds parameter dimensions
}
