% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eg_input.R
\name{fit_ellipsoid_r}
\alias{fit_ellipsoid_r}
\title{Fit an ellipsoid to the noisy data using Cayley transform ellipsoid fitting (CTEF)}
\usage{
fit_ellipsoid_r(X, k)
}
\arguments{
\item{X}{A data set, in the matrix format.}

\item{k}{The ellipsoid dimension (>1).}
}
\value{
The parameters of the fitted ellipsoid.
}
\description{
\code{fit_ellipsoid_r} fits an ellipsoid to the noisy data and outputs relevant parameters
using Cayley transform ellipsoid fitting (CTEF)  \insertCite{melikechi2023ellipsoid}{ellipsoidgaussian}.
}
\examples{
fit_ellipsoid_r(shell, 3)
}
