% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_posterior_predictive.R
\name{gen_posterior_predictive}
\alias{gen_posterior_predictive}
\title{Random generation from the posterior predictive distribution of an ellipsoid-Gaussian likelihood}
\usage{
gen_posterior_predictive(n, samples, burnin, include_noise = TRUE)
}
\arguments{
\item{n}{number of samples to be generated from the posterior predictive
distribution.}

\item{samples}{a list of results, same format as the output of
\code{\link[=ellipsoid_gaussian]{ellipsoid_gaussian()}} or \code{\link[=gSGNHT_EG]{gSGNHT_EG()}}.}

\item{burnin}{number of burn-in samples, that is the number of samples to be
discarded from \code{samples}.}

\item{include_noise}{Logical. If true, the noise (which has covariance
\code{Sigma}) is included; else, it is excluded and only the resulting data
will be on the ellipsoid.}
}
\value{
A matrix of data, samples from the posterior predictive distribution
}
\description{
\code{gen_posterior_predictive} generates from the posterior distribution of an ellipsoid-Gaussian likelihood,
using the posterior samples saved in the output format of \code{\link[=ellipsoid_gaussian]{ellipsoid_gaussian()}}.
}
\details{
The posterior predictive distribution of \eqn{\tilde{x}} given \eqn{\mathbf{X}} is defined by
\deqn{{p(\tilde{x}|\mathbf{X}) = \int_{\Theta}p(\tilde{x}|\theta)p(\theta|\mathbf{X})d\theta}.}
Therefore, to generate from \eqn{p(\tilde{x}|\mathbf{X}),}
we generate from the posterior distribution \eqn{p(\theta|\mathbf{X})} using saved
samples, and then using the generated parameters to simulate from the likelihood
\eqn{p(\tilde{x}|\theta).}
}
\examples{
gen_posterior_predictive(1000, samples, 2500)
}
