% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{geodesicFlow}
\alias{geodesicFlow}
\title{Find geodesic flow}
\usage{
geodesicFlow(epsilon, theta, v, type)
}
\arguments{
\item{epsilon}{XXX}

\item{theta}{XXX}

\item{v}{XXX}

\item{type}{The type of the surface, either 'euclidean', 'sphere' or 'stiefel'.}
}
\description{
\code{geodesicFlow} calculates the geodesic flow
}
