% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{gSGNHT_EG}
\alias{gSGNHT_EG}
\title{Sample from the posterior distribution of an ellipsoid-Gaussian likelihood}
\usage{
gSGNHT_EG(dat, k, init, prior_par, alg_par, niter, updateCenter)
}
\arguments{
\item{dat}{A numeric matrix of data (n by p).}

\item{k}{The latent dimension.}

\item{init}{A list of initial values to be used for the sampler.}

\item{prior_par}{A list of prior parameters}

\item{alg_par}{A list of algorithm-related parameters.}

\item{niter}{The total number of iterations.}

\item{updateCenter}{Logical, whether to update the center or not during sampling.}
}
\value{
A list of samples, including samples for \code{axesDir} (axes directions),
\code{mu} (\eqn{\mu}),\code{center} (\eqn{\mathbf{c}}),
\code{axesLen} (axes lengths), \code{Lambda} (\eqn{\Lambda}), \code{tau} (\eqn{\tau}) and
\code{invsig} (\eqn{1 / \sigma_j^2, j = 1, \ldots, p}). For each parameter,
the samples are saved as either as an array (e.g. \code{Lambda}) or as
a matrix (e.g. \code{axesLen} and \code{tau}), with the first dimension being
the total number of iterations.
}
\description{
\code{gSGNHT_EG} draws from the posterior distribution associated with an ellipsoid-Gaussian
likelihood using geodesic Stochastic Gradient Nose-Hoover Thermostats (gSGNHT)
\insertCite{liu16sgmcmc}{ellipsoidgaussian}.
}
\examples{
  if (reticulate::py_module_available('ctef')) {
 parList <- gen_input_eg(shell,k = 3, updateCenter = TRUE)
 res <- gSGNHT_EG(shell, parList$k, parList$init, parList$prior_par,
                 parList$alg_par, 100, parList$updateCenter)
                 }
}
