% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{gSGNHT_EG_sampling}
\alias{gSGNHT_EG_sampling}
\title{Posterior sampler}
\usage{
gSGNHT_EG_sampling(
  dat,
  latent_dim,
  variable_interest,
  prior_para,
  par_type,
  epsilon1,
  Apara,
  niter,
  minibatchSize = 50
)
}
\arguments{
\item{dat}{The data set, a numeric matrix of N by p.}

\item{latent_dim}{The latent dimension.}

\item{variable_interest}{The variables of interest.}

\item{prior_para}{The list of prior parameters.}

\item{par_type}{The parameter types.}

\item{epsilon1}{The step size used in the gSGNHT sampler.}

\item{Apara}{The scalar parameter used in the gSGNHT sampler.}

\item{niter}{The total number of iterations.}

\item{minibatchSize}{The minibatch size.}
}
\value{
A list of the samples, by names of the parameters.
}
\description{
\code{gSGNHT_EG_sampling} draws posterior samples using the geodesic stochastic
gradient Nose-Hoover Thermostat algorithm.
}
