% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eg_input.R
\name{init_CTEF}
\alias{init_CTEF}
\title{Produce initial values to the sampler}
\usage{
init_CTEF(dat, noise_var, k, ordered)
}
\arguments{
\item{dat}{A numeric matrix of data, N by p.}

\item{noise_var}{A vector of estimates of the noise variances, length = p.}

\item{k}{The latent dimension.}

\item{ordered}{Logical, whether to order the axes lengths in a monotonically
increasing manner. Typically we set it to be false.}
}
\value{
A list of the initial values, ordered by parameter names.
}
\description{
\code{init_CTEF} provides initial values to the sampler using CTEF.
}
