% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{joint_rot_samples}
\alias{joint_rot_samples}
\title{Resolve rotational ambiguity in samples of factor loadings and factors jointly}
\usage{
joint_rot_samples(lambda_samps, lat_facs, iterations)
}
\arguments{
\item{lambda_samps}{The lambda samples in the list of samples as output of
\code{\link[=ellipsoid_gaussian]{ellipsoid_gaussian()}}.}

\item{lat_facs}{The latent factors, an array of samples, number of observations
x latent dimensions x number of samples}

\item{iterations}{Iteration indices of the samples that are used for drawing
the latent factors, a vector.}
}
\description{
\code{joint_rot_samples} performs the varimax rotation on the factor loadings samples
and column-based matching to resolve sign and label switching; see the help page of
\code{\link[infinitefactor:jointRot]{infinitefactor::jointRot()}} for more details.
}
