% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{logDensitylinvSig}
\alias{logDensitylinvSig}
\title{Log-density of log(precision)}
\usage{
logDensitylinvSig(linvSig, asig)
}
\arguments{
\item{linvSig}{log(precision).}

\item{asig}{The standard deviation parameter in the prior distribution.}
}
\value{
The total contribution from the prior on the logarithm scale.
}
\description{
\code{logDensitylinvSig} calculates the log prior density of the log(precision),
which is the parameter that is being updated in the sampler.
}
\details{
Each of the variances, 1 / precision, follows a normal distribution.
}
