% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{logDensityTau_prior}
\alias{logDensityTau_prior}
\title{Evaluate the log density of the precision parameters}
\usage{
logDensityTau_prior(tau_prop, tau_prior)
}
\arguments{
\item{tau_prop}{Proposed tau parameter}

\item{tau_prior}{A list of tau prior parameters, including cvmf (mean),
dvmf (standard deviation), lowerB and upperB.}
}
\description{
\code{logDensitylinvSig} evaluates the log density of the precision parameters.
}
