% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{logit}
\alias{logit}
\title{The logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{The proportion, between 0 and 1.}
}
\description{
\code{logit} calculates \eqn{logit(p) = \log(\frac{p}{1 - p})}.
}
