% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gradient.R
\name{logTransGradAdjust}
\alias{logTransGradAdjust}
\title{Gradient change due to log transformation}
\usage{
logTransGradAdjust(var, orig_deri, lowerB = 0)
}
\arguments{
\item{var}{Variable before taking the logarithm.}

\item{orig_deri}{The derivative w.r.t the variable.}

\item{lowerB}{The lower bound of the variable \code{var}.}
}
\value{
Derivative w.r.t log(\code{var}).
}
\description{
\code{logTransGradAdjust} computes the derivative w.r.t log(\code{var}) given
the derivative w.r.t \code{var}.
}
