% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logTransGradAdjust_cpp}
\alias{logTransGradAdjust_cpp}
\title{Gradient change due to log transformation}
\usage{
logTransGradAdjust_cpp(invSig, invSig_jaco)
}
\arguments{
\item{invSig}{A numeric vector, the parameter values.}

\item{invSig_jaco}{The associated Jacobian.}
}
\description{
\code{logTransGradAdjust_cpp} calculates the gradient change due to log transformation.
}
