% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lowerTriOuterProduct_cpp}
\alias{lowerTriOuterProduct_cpp}
\title{Lower Triangular of the outer product}
\usage{
lowerTriOuterProduct_cpp(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\description{
\code{lowerTriOuterProduct_cpp} calculates the outer product of \code{vec}
and returns the lower triangular part of the resulting matrix
}
