% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{misIdx}
\alias{misIdx}
\title{Missing index of the data entries}
\usage{
misIdx(dat)
}
\arguments{
\item{dat}{a numeric matrix, data.}
}
\description{
\code{misIdx} finds the positions for the missing entries in \code{dat}.
}
