% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mussels.R
\docType{data}
\name{mussels}
\alias{mussels}
\title{Horse mussels data}
\format{
A data frame with 201 rows and 5 variables
\describe{
\item{H}{Shell height in mm}
\item{L}{Shell length in mm}
\item{S}{Shell mass in gram}
\item{W}{Shell width in mm}
\item{M}{Muscle mass in gram, the mass of the edible portion of the mussles.}
}
}
\usage{
data(mussels)
}
\description{
The data were furnished by Mike Camden, Wellington Polytechnic, Wellington,
New Zealand. Horse mussels, Atrinia, were studied from 5 sites in the
Marlborough Sounds in New Zealand in December 1984. The response is muscle
mass, M, the mass of the edible portion of the mussels. Some of the possible
predictors are shell length, width, height and mass.
}
\references{
\insertRef{cook1998regression}{ellipsoidgaussian}
}
\keyword{datasets}
