% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{mycorrelations}
\alias{mycorrelations}
\title{Correlation for the GGally::ggpairs}
\usage{
mycorrelations(data, mapping, ...)
}
\arguments{
\item{data}{A matrix of data.}

\item{mapping}{Mapping.}

\item{...}{...}
}
\description{
\code{mycorrelations} supplies the correlation used for the GGally::ggpairs.
}
