% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{mydensity}
\alias{mydensity}
\title{Density plots for the GGally::ggpairs plot}
\usage{
mydensity(data, mapping, ...)
}
\arguments{
\item{data}{A data frame of data.}

\item{mapping}{Mapping.}

\item{...}{...}
}
\description{
\code{mydensity} provides the density plots used in the GGally::ggpairs plot.
}
