% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{myscatter}
\alias{myscatter}
\title{Scatter plots for the GGally::ggpairs plot}
\usage{
myscatter(data, mapping, ...)
}
\arguments{
\item{data}{A matrix of data.}

\item{mapping}{Mapping.}

\item{...}{...}
}
\description{
\code{myscatter} provides the scatter plots used in the GGally::ggpairs plot.
}
