% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{orthogonalProj}
\alias{orthogonalProj}
\title{Orthogonal projection}
\usage{
orthogonalProj(theta, type)
}
\arguments{
\item{theta}{A vector, the parameter values}

\item{type}{A string, either 'euclidean' or 'sphere'}
}
\description{
\code{orthogonalProj} finds the orthogonal projection of \code{theta} based on
the surface type.
}
