% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{pairsPlot}
\alias{pairsPlot}
\title{ggplot2 pairs plot with two data sets in juxtaposition}
\usage{
pairsPlot(dat_orig, dat_new, dir = NULL, name = NULL)
}
\arguments{
\item{dat_orig}{A matrix of data, e.g. the original data set.}

\item{dat_new}{A matrix of data, e.g. the data generated from the posterior
predictive distribution.}

\item{dir}{A string, a directory where the resulting image is saved.}

\item{name}{A string, the name of the resulting image.}
}
\value{
ggmatrix object from GGally, a matrix of plots
}
\description{
\code{pairsPlot} makes a matrix of plots with two data sets in juxtaposition with
correlation in the upper triangular part, density on the diagonal, and
scatter plots on the lower triangular part.
}
\details{
Arguments \code{dir} and \code{name} provide the option to save
the plot at location \code{dir} with name \code{name}.
}
\examples{
dat_new <- gen_posterior_predictive(1000, samples, 2500)
if (require('GGally') && require('ggplot2') && require('dplyr') &&
require('scales')) {
 pairsPlot(shell, dat_new)
  }
}
