% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{pairsPlot_onedata}
\alias{pairsPlot_onedata}
\title{ggplot2 pairs plot with correlation information}
\usage{
pairsPlot_onedata(dat_new, dir = NULL, name, rotate_x_text = F)
}
\arguments{
\item{dat_new}{A matrix of data.}

\item{dir}{String, the path to where the plot should be saved. Default is NULL, meaning
the plot is not saved.}

\item{name}{String, the name of the plot to be saved. When \code{dir} is NULL, meaning
the plot will not be saved, the name can be left blank.}

\item{rotate_x_text}{Logical, whether to rotate the texts on the x-axis or not.
If so, they are positioned at a 60-degree angle.}
}
\value{
ggmatrix object from GGally, a matrix of plots
}
\description{
\code{pairsPlot_onedata} makes a matrix of plots for a given data set with correlation in the upper
triangular part, density on the diagonal, and scatter plots on the lower triangular part.
}
\details{
Arguments \code{dir} and \code{name} provide the option to save
the plot at location \code{dir} with name \code{name}.
\code{rotate_x_text} also allows users to rotate the texts on the
x-axis in case of space constraint.
}
\examples{
pairsPlot_onedata(shell, NULL)
}
