% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{plot_factor_loadings}
\alias{plot_factor_loadings}
\title{Visualize the factor loadings}
\usage{
plot_factor_loadings(lambda_samples, row_idx = NULL)
}
\arguments{
\item{lambda_samples}{A list of the samples, length = number of samples,
each element is p by k.}

\item{row_idx}{A vector row indices of the factor loadings to be plotted. Default
is all the rows are included.}
}
\description{
\code{plot_factor_loadings} visualizes the posterior mean of the factor loadings matrix. We
recommend post processing the samples first using \code{\link[=draw_latent_factors]{draw_latent_factors()}} and
\code{\link[=postprocess]{postprocess()}}.
}
