% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{plotmat}
\alias{plotmat}
\title{Visualize the factor loadings}
\usage{
plotmat(mat, color = "green", title = NULL, args = NULL, saveDir = NULL)
}
\arguments{
\item{mat}{A matrix of numeric values.}

\item{color}{The color scheme of the plot, "green", "red" or "wes".}

\item{title}{A string, optional plot title.}

\item{args}{Optional additional ggplots arguments}

\item{saveDir}{The path to where the image is saved. Default is NULL, meaning
the figure is not saved. When it is saved, it is named as loadings.pdf.}
}
\description{
\code{plotmat} builds on the function \code{\link[infinitefactor:plotmat]{infinitefactor::plotmat()}} and includes
additional features, including an option to save the figure and automatic
inclusion of the variable names in the plot.
}
\examples{
mat <- matrix(rnorm(6), 3, 2)
plotmat(mat = mat)

}
