% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{predictY}
\alias{predictY}
\title{Impute the mising values}
\usage{
predictY(temp, tau_par, dat, missing_idx, mis_pos)
}
\arguments{
\item{temp}{A list of the distribution parameters.}

\item{tau_par}{A list of parameters regarding tau.}

\item{dat}{A numeric matrix, data.}

\item{missing_idx}{Indices of the missing entries.}

\item{mis_pos}{Positions of the missing entries.}
}
\description{
\code{predictY} imputes the missing entries in \code{dat} by assuming that
the data comes from an ellipsoid-Gaussian distribution.
}
