% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eg_input.R
\name{preprocess_Lambda_ctef}
\alias{preprocess_Lambda_ctef}
\title{Process Lambda}
\usage{
preprocess_Lambda_ctef(Lambda, ax_lengths)
}
\arguments{
\item{Lambda}{The factor loading matrix output from CTEF.}

\item{ax_lengths}{The axes lengths output from CTEF.}
}
\value{
A list of axes lengths and directions.
}
\description{
\code{preprocess_Lambda_ctef} processes the factor loading matrix and
the axes length output from \code{\link[=init_CTEF]{init_CTEF()}} such that the order of the
axes lengths match that of the singular values of the factor loading matrix.
}
