% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{projectionTanget}
\alias{projectionTanget}
\title{Projection onto the tangent space}
\usage{
projectionTanget(cur_pos, tobeprojected, type)
}
\arguments{
\item{cur_pos}{Current positions.}

\item{tobeprojected}{The surface to be projected.}

\item{type}{A string, either sphere or stiefel.}
}
\description{
\code{projectionTanget} finds the orthogonal projection of \code{tobeprojected} based
on the surface type.
}
