% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{projectionTanget_stiefel}
\alias{projectionTanget_stiefel}
\title{projection on the tangent of the stiefel manifold}
\usage{
projectionTanget_stiefel(U, X)
}
\arguments{
\item{U}{R_d by p}

\item{X}{V_d by p}
}
\description{
\code{projectionTanget_stiefel} projects on the tangent surface of the
stiefel manifold
}
