% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{proposeSigTau_ram2}
\alias{proposeSigTau_ram2}
\title{Adaptive Metropolis update for the precision and tau.}
\usage{
proposeSigTau_ram2(all_theta, prior_para, iteration, SS, dat, case)
}
\arguments{
\item{all_theta}{A list of the parameters}

\item{prior_para}{A list of prior parameters}

\item{iteration}{the current iteration index}

\item{SS}{current covariance matrix}

\item{dat}{A numeric matrix, the entire data matrix}

\item{case}{If case is 2, both tau and variances are updated; if case is 1,
only variances are updated. Otherwise, neither parameters are updated.}
}
\description{
\code{proposeSigTau_ram2} uses adaptive Metropolis (Vihola 2010)
to update the noise variances and tau.
}
\references{
Vihola, M. (2010). Robust adaptive Metropolis algorithm with coerced
acceptance rate. Statistics and Computing, 22, 997-1008.
}
