% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{randomDrift_fast}
\alias{randomDrift_fast}
\title{Draw the random drift in Step O of the sampling procedure}
\usage{
randomDrift_fast(Apara, epsilon, para_dim, Bmat = 0)
}
\arguments{
\item{Apara}{The scalar parameter}

\item{epsilon}{step size}

\item{para_dim}{The parameter dimensions}

\item{Bmat}{V(x^*) in step O}
}
\value{
A list of vectors, with each of length equal to the parameter
dimension.
}
\description{
\code{randomDrift_fast} draws the random drift from a Gaussian distribution in step
O of the sampling procedure
}
