% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_density_eg.R
\name{rellipsoidgaussian}
\alias{rellipsoidgaussian}
\title{Density and random generation for the ellipsoid-Gaussian distribution}
\usage{
rellipsoidgaussian(n, center, Lambda, mu, tau, Sigma)
}
\arguments{
\item{n}{Number of random vectors to be simulated.}

\item{center}{Vector of length \eqn{p}, representing the center of the underlying ellipsoid.}

\item{Lambda}{Matrix of dimension \eqn{p \times k}, representing the ellipsoid shape or the factor loading matrix.
in the von-Mises Fisher linear factor model.}

\item{mu}{Vector of length \eqn{k}.}

\item{tau}{The concentration parameter, non-negative.}

\item{Sigma}{Covariance matrix of the noise (\eqn{p \times p}), typically diagonal.}
}
\value{
n samples from the ellipsoid-Gaussian distribution.
}
\description{
Density and random generation for the multivariate ellipsoid-Gaussian distribution
with parameters \code{center}, \code{Lambda}, \code{mu}, \code{tau} and \code{Sigma}.
}
\details{
An ellipsoid-Gaussian distribution with parameters \eqn{\mathbf{c}} (\code{center}), \eqn{\Lambda} (\code{Lambda}),
\eqn{\mu} (\code{mu}), \eqn{\tau} (\code{tau}) and \eqn{\Sigma} (\code{Sigma})
has density
\deqn{f(\mathbf{x}) = \frac{C_k(\tau)}{(2\pi)^{\frac{p}{2}}\prod_{i = 1}^p \sigma_i}
\exp\left\{-\frac{1}{2} (x - {c})^T \Sigma^{-1} (x - {c})\right\}
\varsigma\left\{ \tau \mu + \Lambda^T \Sigma^{-1
}(x - {c}) , \frac{\Lambda^T \Sigma^{-1}\Lambda}{2}\right\},}
where \eqn{C_k(\tau) = ({\tau}/{2})^{k / 2 - 1}\{\Gamma(k / 2)I_{k /2 - 1}(\tau)\}^{-1}} and \eqn{I_{v}(\cdot)}
denotes the modified Bessel function of the first kind of order \eqn{v}, with respect to the Lebesgue measure on \eqn{\mathbb{R}^p,}
\eqn{\varsigma(\vartheta, A)} is the normalizing constant in a Fisher-Bingham distribution with density
\deqn{\frac{1}{\varsigma(\kappa \vartheta, A)} \exp\left(\kappa \vartheta^T{y} - {y}^T A {y}\right).} It is generated by
marginalizing over the distribution of \eqn{\eta} in a von-Mises Fisher linear factor model defined as follows:
\deqn{\mathbf{x} = \mathbf{c} + \Lambda \eta + \epsilon,\quad \epsilon \sim \text{N}_p({0}, \Sigma), \quad
\eta \sim \text{vMF}(\mu, \tau) \quad
\text{and} \quad  \Sigma = \text{diag}\left(\sigma_1^2, \ldots, \sigma_p^2\right).}
}
\examples{
dat <- rellipsoidgaussian(1000, rep(0,3), matrix(rnorm(6, mean = -5, sd = 2),nrow = 3, ncol = 2),
c(0,1), 1, diag(0.5,3))
}
