% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reorder_vec}
\alias{reorder_vec}
\title{Reorder vector}
\usage{
reorder_vec(v, idx)
}
\arguments{
\item{v}{A vector.}

\item{idx}{A vector of indices.}
}
\description{
\code{reorder_vec} orders vector \code{v} based on the order \code{idx}.
}
