% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{rFB_MH}
\alias{rFB_MH}
\title{Simulation from Fisher-Bingham distribution}
\usage{
rFB_MH(eta1, para1, A)
}
\arguments{
\item{eta1}{A vector of length k}

\item{para1}{A vector of length k}

\item{A}{A square matrix, k by k.}
}
\description{
\code{rFB_MH} simulates from a Fisher-Bingham distribution using
Hoff (2009).
}
\references{
Peter D. Hoff (2009) Simulation of the Matrix Bingham–von Mises–Fisher
Distribution, With Applications to Multivariate and Relational Data,
Journal of Computational and Graphical Statistics, 18:2, 438-456,
DOI: 10.1198/jcgs.2009.07177
}
