% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{root4SaddleEquat}
\alias{root4SaddleEquat}
\title{Find the root to the saddle point equation.}
\usage{
root4SaddleEquat(gam, lam, low, up)
}
\arguments{
\item{gam}{A vector of length n.}

\item{lam}{A vector of length n.}

\item{low}{The lower bound of the solution.}

\item{up}{The upper bound of the solution.}
}
\description{
\code{root4SaddleEquat} finds the root to the saddlepoint equation
}
