% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotate_cpp}
\alias{rotate_cpp}
\title{The rotation from a to b.}
\usage{
rotate_cpp(a, b)
}
\arguments{
\item{a}{A vector of length n.}

\item{b}{A vector of length n.}
}
\description{
\code{rotate_cpp} finds the rotation matrix that transforms a to b.
}
