% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{saddle_equat_cpp}
\alias{saddle_equat_cpp}
\title{Find saddle point}
\usage{
saddle_equat_cpp(ta, gam, lam)
}
\arguments{
\item{ta}{t.}

\item{gam}{Gamma.}

\item{lam}{Lambda.}
}
\description{
\code{saddle_equat_cpp} is a helper function for evaluating the pseudo-normalising constant.
See \insertCite{KumeWood05}{ellipsoidgaussian} for more details.
}
\references{
\insertAllCited{}
}
