% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\docType{data}
\name{shell}
\alias{shell}
\title{A data set simulated from the ellipsoid-Gaussian distribution}
\format{
An object of class \code{"matrix"}
}
\usage{
data(shell)
}
\description{
A small data set that is generated from the ellipsoid-Gaussian distribution is
included to illustrate the use of the package. The shape of the data set resembles
a shell and thus the name.
}
\examples{

data(shell)
head(shell)
pairsPlot_onedata(shell)

}
\references{
This data set was artificially created for the ellipsoidgaussian package.
}
\keyword{datasets}
