% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{simpleImpute}
\alias{simpleImpute}
\title{Impute missing data}
\usage{
simpleImpute(dat)
}
\arguments{
\item{dat}{A numerical matrix of p by k, with missing entries coded as NA.}
}
\value{
Imputed data, a numerical matrix of p by k.
}
\description{
\code{simpleImpute} imputes missing data using \code{\link[mice:mice]{mice::mice()}}.
}
