% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{storeUpdates}
\alias{storeUpdates}
\title{Format the storage for the gSGNHT output}
\usage{
storeUpdates(para_dims, niter, init_vals)
}
\arguments{
\item{para_dims}{Dimensions of parameters}

\item{niter}{Total number of iterations}

\item{init_vals}{A list of initial values by parameter names}
}
\value{
A list for storing output of gSGNHT sampler
}
\description{
\code{storeUpdates} creates a list by parameter names and initializes
a storage container for each parameter.
}
