% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{transformedTau2Tau}
\alias{transformedTau2Tau}
\title{Transform tau}
\usage{
transformedTau2Tau(par, lowerB, upperB, fac)
}
\arguments{
\item{par}{logit((tau - l) / (u - l)).}

\item{lowerB}{The lower bound of tau.}

\item{upperB}{The upper bound of tau.}

\item{fac}{Default is 1.}
}
\description{
\code{transformedTau2Tau} transforms logit((tau - l) / (u - l)), the parameter
that is being updated to tau, where u and l are the upper and lower bounds of
tau respectively.
}
