% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Lambda_prior_pars_optional.R
\name{updateKap}
\alias{updateKap}
\title{Update Kap, a parameter in the hyperprior of Lambda}
\usage{
updateKap(a, phi1, lambda1)
}
\description{
Update Kap, a parameter in the hyperprior of Lambda
}
