% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Lambda_prior_pars_optional.R
\name{updatePhi}
\alias{updatePhi}
\title{Update Phi, a parameter in the hyperprior of Lambda}
\usage{
updatePhi(a, lambda1)
}
\description{
Update Phi, a parameter in the hyperprior of Lambda
}
