% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{updateSample_gSGNHT}
\alias{updateSample_gSGNHT}
\title{Sampling procedure for one iteration of update}
\usage{
updateSample_gSGNHT(
  all_theta,
  par_type,
  para_dims,
  para_dim_updated,
  aux_par,
  xi,
  dat,
  k,
  epsilon,
  Apara,
  prior_para,
  accurate_grad,
  minibatchSize,
  initialization_period,
  SS = NULL,
  iteration = NULL,
  Lsteps = 10
)
}
\arguments{
\item{all_theta}{A list of the parameter values}

\item{par_type}{Parameter types}

\item{para_dims}{Parameter dimensions}

\item{para_dim_updated}{The dimensions of the parameters being updated.}

\item{aux_par}{auxiliary parameters}

\item{xi}{The thermostat parameter}

\item{dat}{A numerical matrix of p by k, the data}

\item{k}{The latent dimension.}

\item{epsilon}{step size}

\item{Apara}{the scalar parameter}

\item{prior_para}{A list of prior parameters}

\item{accurate_grad}{Logical, whether to use accurate or simple gradient.}

\item{minibatchSize}{Minibatch size.}

\item{initialization_period}{If NULL, then not initialization period. Upweighing
certain observations is disabled.}

\item{SS}{The covariance matrix parameter used in \code{\link[=proposeSigTau_ram2]{proposeSigTau_ram2()}}.}

\item{iteration}{Total number of iterations.}

\item{Lsteps}{Number of the repetition of the ABOBA steps. Default is 10.}
}
\description{
\code{updateSample_gSGNHT} is the sampling procedure for one iteration of update
in gSGNHT.
See Algorithm 2 in the Appendix of Liu et al. (2016)
for more details.
}
\references{
Liu, C., Zhu, J., and Song, Y. (2016). Stochastic gradient geodesic MCMC
methods. In Advances in Neural Information Processing Systems 29, pages
3009–3017.
}
