% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{updateSample_stepA}
\alias{updateSample_stepA}
\title{Update the sample in step A}
\usage{
updateSample_stepA(
  par_name,
  para_dim,
  par_type,
  all_theta,
  aux_par,
  xi,
  epsilon
)
}
\arguments{
\item{par_name}{parameter names}

\item{para_dim}{parameter dimensions}

\item{par_type}{parameter types (in Euclidean space or not)}

\item{all_theta}{all the parameter values, saved in a list}

\item{aux_par}{auxiliary parameter V}

\item{xi}{The thermostats parameter}

\item{epsilon}{step size}
}
\description{
\code{updateSample_stepA} updates the samples in step A of the gSGNHT sampling
procedure in Algorithm 2  in the Appendix of \insertCite{liu16sgmcmc;textual}{ellipsoidgaussian}.
}
