% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{updateV_stepB}
\alias{updateV_stepB}
\title{Step B in the gSGNHT sampler}
\usage{
updateV_stepB(epsilon, xi, v)
}
\arguments{
\item{epsilon}{The step size.}

\item{xi}{XXX.}

\item{v}{A list of the auxiliary parameters.}
}
\value{
A list of the updated auxiliary parameters.
}
\description{
\code{updateV_stepB} updates the step B in the sampling procedure of gSGNHT.
See Algorithm 2 in the Appendix of \insertCite{liu16sgmcmc;textual}{ellipsoidgaussian}.
}
\references{
\insertAllCited{}
}
