% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{updateV_stepB_ctrlCov}
\alias{updateV_stepB_ctrlCov}
\title{step B in the gSGNHT sampler}
\usage{
updateV_stepB_ctrlCov(epsilon, xi, v, covContrlPar = NULL)
}
\arguments{
\item{epsilon}{Step size.}

\item{xi}{Thermostat.}

\item{v}{The auxiliary parameter.}

\item{covContrlPar}{If not controlling covariance, then set to NULL; else,
covariance control parameter.}
}
\value{
A list of updated V
}
\description{
\code{updateV_stepB_ctrlCov} runs step B in the sampling procedure with
controlled variance option. The controlled variance option is currently
disabled in the sampler since it does not help much with the performance.
}
