% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_eg_sampler.R
\name{updateV_stepO_fast}
\alias{updateV_stepO_fast}
\title{Update V in the gSGNHT sampling step O}
\usage{
updateV_stepO_fast(
  dat,
  para_dim,
  k,
  par_notin_euclidean,
  all_theta,
  v,
  epsilon,
  Apara,
  prior_para,
  accurate_grad,
  minibatchSize,
  initialization_period
)
}
\arguments{
\item{dat}{The minibatch, a numerical matrix of n by p.}

\item{para_dim}{Parameter dimensions.}

\item{k}{The latent dimension.}

\item{par_notin_euclidean}{Parameters that are not in the Euclidean space,
e.g. mu.}

\item{all_theta}{A list of the parameter values.}

\item{v}{The auxiliary parameter.}

\item{epsilon}{The step size.}

\item{Apara}{The scalar parameter.}

\item{prior_para}{A list of the prior parameters.}

\item{accurate_grad}{Logical, whether to use a simple or accurate gradient.}

\item{minibatchSize}{minibatch size.}

\item{initialization_period}{If NULL, then not initialization period.}
}
\value{
A list of updated V and gradient
}
\description{
\code{updateV_step0_fast} updates the auxiliary parameter V in step O of the
sampling procedure. See Algorithm 2 in the Appendix of Liu et al. (2016)
for more details.
}
\references{
Liu, C., Zhu, J., and Song, Y. (2016). Stochastic gradient geodesic MCMC
methods. In Advances in Neural Information Processing Systems 29, pages
3009–3017.
}
