% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Vech2Mat_cpp}
\alias{Vech2Mat_cpp}
\title{Convert a vech format to a matrix}
\usage{
Vech2Mat_cpp(para2_vech)
}
\arguments{
\item{para2_vech}{A vector obtained through vech operation.}
}
\description{
\code{Vech2Mat_cpp} converts a single vector of length \eqn{d(d+1)/2} obtained
through the vech (vector half) operation to its original matrix format. The vech
(vector half) operator takes a symmetric \eqn{d \times d} matrix and
stacks the lower triangular half into a single vector of length \eqn{d(d+1)/2}.
}
