% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{approxFBconst_cpp}
\alias{approxFBconst_cpp}
\title{Fisher-Bingham normalising constant}
\usage{
approxFBconst_cpp(para1, para2, idx_)
}
\arguments{
\item{para1}{The vector parameter, length is k.}

\item{para2}{The matrix parameter, k by k.}

\item{idx_}{The order of approximation,1: First order; 2: second order first
type; 3: second order second type.}
}
\description{
\code{approxFBconst_cpp} approximates the normalising constant in the Fisher-Bingham
distribution using saddlepoint approximation \insertCite{KumeWood05}{ellipsoidgaussian}.
}
\references{
\insertAllCited{}
}
