% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{draw_latent_factors}
\alias{draw_latent_factors}
\title{Sample latent factors}
\usage{
draw_latent_factors(
  dat,
  samples,
  burnin = NULL,
  num_samp = 200,
  dir2Save = NULL
)
}
\arguments{
\item{dat}{A matrix of data, n by p.}

\item{samples}{A list of posterior samples, output of \code{\link[=ellipsoid_gaussian]{ellipsoid_gaussian()}}.}

\item{burnin}{Number of burn-in samples.}

\item{num_samp}{Number of samples for each latent factor.}

\item{dir2Save}{A string, the path to where the sampled latent factors should
be saved. Default is NULL, which means the latent factors are not saved.}
}
\description{
\code{draw_latent_factors} simulates \eqn{\min(500,
\text{number of post burn-in samples})} latent factors for each observation from
their posterior distribution, a Fisher-Bingham distribution. Simulating from
a Fisher-Bingham is challenging and we use the method by Hoff (2009).
}
\examples{
\dontrun{
lat_facs <- draw_latent_factors(shell, samples, burnin = 2500, num_samp = 5)
}

}
\references{
Peter D. Hoff (2009) Simulation of the Matrix Bingham–von Mises–Fisher
Distribution, With Applications to Multivariate and Relational Data,
Journal of Computational and Graphical Statistics, 18:2, 438-456,
DOI: 10.1198/jcgs.2009.07177
}
