% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfb_cpp}
\alias{kfb_cpp}
\title{Helper function for evaluating the pseudo-normalising constant}
\usage{
kfb_cpp(j, gam, lam, ta)
}
\arguments{
\item{j}{The power}

\item{gam}{Gamma.}

\item{lam}{Lambda.}

\item{ta}{t.}
}
\description{
\code{kfb_cpp} is a helper function for evaluating the pseudo-normalising constant.
See \insertCite{KumeWood05}{ellipsoidgaussian} for more details.
}
\references{
\insertAllCited{}
}
