% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_result.R
\name{plot_factor_loadings}
\alias{plot_factor_loadings}
\title{Visualize the factor loadings}
\usage{
plot_factor_loadings(
  lambda_samples,
  variable_names = NULL,
  row_idx = NULL,
  saveDir = NULL
)
}
\arguments{
\item{lambda_samples}{A list of the samples, length = number of samples,
each element is p by k.}

\item{variable_names}{A vector of strings, length = p, variable names. Default
is NULL, meaning unspecified.}

\item{row_idx}{A vector row indices of the factor loadings to be plotted. Default
is all the rows are included.}

\item{saveDir}{A string, the path to where the image should be saved. Default
is NULL, meaning the image is not saved. When the image is saved, its name is
loadings.pdf.}
}
\description{
\code{plot_factor_loadings} visualizes the posterior mean of the factor loadings matrix. We
recommend post processing the samples first using \code{\link[=draw_latent_factors]{draw_latent_factors()}} and
\code{\link[=postprocess]{postprocess()}}. The function uses (post-processed) samples of factor loadings
directly.
}
\examples{
lambda_samples <- list(a = matrix(rnorm(6),3,2), b = matrix(rnorm(6),3,2))
plot_factor_loadings(lambda_samples = lambda_samples)
}
