% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_latent_factor.R
\name{postprocess}
\alias{postprocess}
\title{Post-process the factor loadings}
\usage{
postprocess(dat, samples, burnin = NULL, num_samps = 200, dir2Save = NULL)
}
\arguments{
\item{dat}{A matrix of data, n by p.}

\item{samples}{A list of samples from the posterior distribution, output of
\code{\link[=ellipsoid_gaussian]{ellipsoid_gaussian()}}.}

\item{burnin}{Number of burn-in samples.}

\item{num_samps}{Number of samples for each latent factor.}

\item{dir2Save}{A string, the path to where the latent factors and processed
samples are saved.}
}
\description{
\code{postprocess} post process the samples of the factor loadings to resolve
rotational ambiguity. It performs the varimax rotation on the factor loadings samples
and column-based matching to resolve sign and label switching; see the help page of
\code{\link[infinitefactor:jointRot]{infinitefactor::jointRot()}} for more details.
}
\examples{
\dontrun{
aligned <- postprocess(shell, samples, burnin = 2500, num_samps = 5)
}

}
