% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleLatentFac.R
\name{rFB_MH}
\alias{rFB_MH}
\title{Simulation from Fisher-Bingham distribution}
\usage{
rFB_MH(eta1, para1, A)
}
\arguments{
\item{eta1}{A vector of length k}

\item{para1}{A vector of length k}

\item{A}{A square matrix, k by k.}
}
\description{
\code{rFB_MH} simulates from a Fisher-Bingham distribution using
\insertCite{Hoff09FB}{ellipsoidgaussian}.
}
\references{
\insertAllCited{}
}
