% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Lambda_prior_pars_optional.R
\name{updateKap}
\alias{updateKap}
\title{Update Kap, a parameter in the hyperprior of Lambda}
\usage{
updateKap(a, phi1, lambda1)
}
\arguments{
\item{a}{The parameter in the Dirichlet-Laplace prior for Lambda.}

\item{phi1}{The parameter phi in the Dirichlet-Laplace prior.}

\item{lambda1}{A matrix of p by k, the factor loading matrix.}
}
\value{
Updated Kap parameter
}
\description{
\code{updateKap} updates the kappa parameter in the Dirichlet-Laplace prior of Lambda.
}
