% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Lambda_prior_pars_optional.R
\name{updatePhi}
\alias{updatePhi}
\title{Update Phi, a parameter in the hyperprior of Lambda}
\usage{
updatePhi(a, lambda1)
}
\arguments{
\item{a}{The parameter in the Dirichlet-Laplace prior for Lambda,
typically set at 0.5}

\item{lambda1}{A matrix of p by k, the factor loading matrix}
}
\value{
updated Phi parameter
}
\description{
\code{updatePhi} updates Phi in the Dirichlet-Laplace prior of Lambda.
}
