% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Lambda_prior_pars_optional.R
\name{updatePsi}
\alias{updatePsi}
\title{Update Psi, a parameter in the hyperprior of Lambda}
\usage{
updatePsi(kappa1, phi1, lambda1)
}
\description{
Update Psi, a parameter in the hyperprior of Lambda
}
